+++ src/cryptography/hazmat/primitives/hashes.py
@@ -149,11 +149,27 @@ class RIPEMD160(object):
     block_size = 64
 
 
+RIPEMD160 = utils.deprecated(
+    RIPEMD160,
+    __name__,
+    "The RIPEMD160 hash was deprecated in version 1.9.",
+    utils.DeprecatedIn19
+)
+
+
 @utils.register_interface(HashAlgorithm)
 class Whirlpool(object):
     name = "whirlpool"
     digest_size = 64
     block_size = 64
+
+
+Whirlpool = utils.deprecated(
+    Whirlpool,
+    __name__,
+    "The Whirlpool hash was deprecated in version 1.9.",
+    utils.DeprecatedIn19
+)
 
 
 @utils.register_interface(HashAlgorithm)